Add: lecture 2 redux-toolkit
