+++ src/RegionsContainer.jsx
@@ -4,7 +4,7 @@ import styled from '@emotion/styled';
 import {
   selectRegion,
   loadRestaurants,
-} from './actions';
+} from './slice';
 
 import MenuList from './MenuList';
 import MenuItem from './MenuItem';
