+++ src/store.js
@@ -1,9 +1,7 @@
-import { createStore, applyMiddleware } from 'redux';
+import { createStore } from 'redux';
 
-import thunk from 'redux-thunk';
+import reducer from './slice';
 
-import reducer from './reducer';
-
-const store = createStore(reducer, applyMiddleware(thunk));
+const store = createStore(reducer);
 
 export default store;
